+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [In Development] - Unreleased
 
+### Changed
+
+- Chat list parser re-enabled. See Warning in [README.md](README.md).
+
 
 ## [0.0.1-alpha.11] - 2023-08-23